Installation Steps Asterisk in CentOS 7
A brief introduction to the installation of Asterisk in CentOS 7

Asterisk
-----------------------------------------------------------
Asterisk is a free and open source framework for building communications applications and is sponsored by Digium.

Asterisk is a software implementation of a telephone private branch exchange (PBX). It allows telephones interfaced with a variety of hardware technologies to make calls to one another, and to connect to telephony services, such as the public switched telephone network (PSTN) and voice over Internet Protocol (VoIP) services. Its name comes from the asterisk symbol "*".